Adding content to About page. Details below!
Changes:

- Fixed typo in project checklist

- Added photo grid to About page (HTML + CSS) (thank you for the helpful guide!!)

Question: how do you change the photos on each grid card? I see how to change the photo for /every/ grid card, but what if you want a different image on each one?

I won't be able to work much tomorrow or Friday, but I think I'll be able to knock out the rest on Saturday during our coding session! If not, the rest of my weekend is wide open to work on it as well :)