Compressed Images
All images have been compressed. Season 1 & 3 on the about page did not have images that I could replace or link to so I wrote in the images to the cards with the link to the correctly compressed images. Converted images into Browser JPEG, only 3 had to remain PNG in order to maintain the integrity of the site. I cleaned up the images folder so that the images in there are the ones that are being used. Crossing my fingers that i did this correctly and didn't accidentally create more work.